Initial File Dump
This is a random role selecting machine. The program creates Scientist objects and stores them in an array. ("roster") Each Scientist object has (<int>role, <str>name, <int>[] ). The roster is shuffled several times before it is referenced again. Once it is shuffled there are 2 routines that assign the role <int> to each Scientists. It's basically two different filters. 

Routine #1 assigns the first 9 slots of the #OVERALL ROSTER, Line 27. The first role sought out by the machine is Project Manager and down from there. A random Scientist is selected to see if this job was there top pick. If that person is found they are assigned the role. If not, the search continues for 2nd place picks, and on. If none are found then the program will force a decision. 

This allows for greater than %50 of Scientists to get their number one pick. 
(math help)

Routine #2
Once Management and first pick of engi, sci and admin have been set. The rest of the team roles will be assigned in descending order with higher picks getting the roles at random and forced decisions if there are no selections made. 

Print
Line 221 is the goal to aim for. 


State of machine: 
The logic of the machine is %90 there with some implementation challenges.
Last fail: